UX-641 Convert Picture to typescript
